Made changes requested
