+++ src/components/Button.js
@@ -1,0 +1,28 @@
+import React from 'react';
+
+const Buttons = () => (
+  <div className="calculator">
+    <span className="button output">0</span>
+    <span className="button">AC</span>
+    <span className="button">+/-</span>
+    <span className="button">%</span>
+    <span className="button">&divide;</span>
+    <span className="button">7</span>
+    <span className="button">8</span>
+    <span className="button">9</span>
+    <span className="button">&times;</span>
+    <span className="button">4</span>
+    <span className="button">5</span>
+    <span className="button">6</span>
+    <span className="button">&minus;</span>
+    <span className="button">1</span>
+    <span className="button">2</span>
+    <span className="button">3</span>
+    <span className="button">+</span>
+    <span className="button zero">0</span>
+    <span className="button">.</span>
+    <span className="button orange">=</span>
+  </div>
+);
+
+export default Buttons;
