+++ src/components/calculator.js
@@ -1,42 +1,9 @@
 import React from 'react';
 import './calculator.css';
+import Buttons from './Button';
 
-const Buttons = () => (
-  <div className="calculator">
-    <span className="button output">0</span>
-    <span className="button">AC</span>
-    <span className="button">+/-</span>
-    <span className="button">%</span>
-    <span className="button">&divide;</span>
-    <span className="button">7</span>
-    <span className="button">8</span>
-    <span className="button">9</span>
-    <span className="button">&times;</span>
-    <span className="button">4</span>
-    <span className="button">5</span>
-    <span className="button">6</span>
-    <span className="button">&minus;</span>
-    <span className="button">1</span>
-    <span className="button">2</span>
-    <span className="button">3</span>
-    <span className="button">+</span>
-    <span className="button zero">0</span>
-    <span className="button">.</span>
-    <span className="button orange">=</span>
-  </div>
-);
-
-class Calculator extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
-  render() {
-    return (
-      <Buttons />
-    );
-  }
+export default function Calculator() {
+  return (
+    <Buttons />
+  );
 }
-
-export default Calculator;
